requirements.txt
--pre
azure-ai-projects
azure-identity
openai
python-dotenv